requirements.txt
loguru
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
